fetch user component dont delete data on userprofile page working on using comments to redesign
