+++ we-rise/src/app/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import authenticationReducer from '../features/Authentication/AuthenticationSlice';
-import workshopFeedReducer from '../features/WorkshopFeed/WorkshopFeedSlice'
+import workshopFeedReducer from '../features/WorkshopFeed/workshopFeedSlice'
 
 export default configureStore({
   reducer: {
